fetching workings linked to clients.

+++ backend/routes/programs.js
@@ -4,13 +4,13 @@ const {authUser} = require("../middleware/check-auth")
 const router = express.Router();
 
 const Program = require("../models/program");
+const User = require("../models/user")
 
 router.post("",authUser,(req, res, next) => {
     const program = new Program({
       name: req.body.name,
       description: req.body.description,
-      // workouts:
-      // client: req.userData.userId
+      creator: req.userData.userId
     });
 
     program.save().then((createdProgram) => {
@@ -27,10 +27,9 @@ router.put("/:id",authUser,(req, res, next) => {
     _id: req.body.id,
     name: req.body.name,
     description: req.body.description,
-    // workout:
-    // client: req.userData.userId
+    creator: req.userData.userId
   });
-  Program.updateOne({ _id: req.params.id }, program).then((result) => {
+  Program.updateOne({ _id: req.params.id, creator: req.userData.userId }, program).then((result) => {
     if(result.nModified > 0) {
     res.status(200).json({ message: "Update Successful" });
     } else {
@@ -42,14 +41,35 @@ router.put("/:id",authUser,(req, res, next) => {
 router.get("", (req, res, next) => {
  Program.find().then((program)=> {
    if(program){
+    // const creatorName = User.findById()
      res.status(200).json({
        message: "Programs found",
        programs: program
      })
    }
  })
-
 });
+// router.get("", (req, res, next) => {
+//   Program.find().then((program)=> {
+//     if (program){
+//       program.forEach(item => {
+//         item.creator
+//       })
+
+//         console.error(user)
+//         res.status(200).json({
+//           message: "Programs found",
+//           programs: {
+//             program: program,
+//             user: user
+//           }
+//         })
+
+//     }
+//   })
+// })
+
+
 
 router.get("/:id", (req, res, next) => {
   Program.findById(req.params.id).then((program) => {
@@ -62,7 +82,7 @@ router.get("/:id", (req, res, next) => {
 });
 
 router.delete("/:id", authUser, (req, res, next) => {
-  Program.deleteOne({ _id: req.params.id}).then((result) => {
+  Program.deleteOne({ _id: req.params.id, creator: req.userData.userId}).then((result) => {
     if(result.n > 0) {
       res.status(200).json({ message: "Deletion Successful" });
       } else {
